Create custom login example, refactor common bits
