+++ common/sample-web-server.js
@@ -1,0 +1,83 @@
+/**
+ * A simple web server that initializes the OIDC Middleware library with the
+ * given options, and attaches route handlers for the example profile page
+ * and logout functionality.
+ */
+
+const cons = require('consolidate');
+const express = require('express');
+const session = require('express-session');
+const path = require('path');
+const { ExpressOIDC } = require('@okta/oidc-middleware');
+
+const templateDir = path.join(__dirname, '..', 'common', 'views');
+const frontendDir = path.join(__dirname, '..', 'common', 'assets');
+
+module.exports = function SampleWebServer(sampleConfig, extraOidcOptions) {
+
+  const oidc = new ExpressOIDC(Object.assign({
+    issuer: sampleConfig.oidc.issuer,
+    client_id: sampleConfig.oidc.clientId,
+    client_secret: sampleConfig.oidc.clientSecret,
+    redirect_uri: sampleConfig.oidc.redirectUri,
+    scope: sampleConfig.oidc.scope
+  }, extraOidcOptions || {}));
+
+  const app = express();
+
+  app.use(session({
+    secret: 'this-should-be-very-random',
+    resave: true,
+    saveUninitialized: false
+  }));
+
+  // Provide the configuration to the view layer because we show it on the homepage
+  const displayConfig = Object.assign(
+    {},
+    sampleConfig.oidc,
+    {
+      clientSecret: '****' + sampleConfig.oidc.clientSecret.substr(sampleConfig.oidc.clientSecret.length - 4, 4)
+    }
+  );
+
+  app.locals.oidcConfig = displayConfig;
+
+  // This server uses mustache templates located in views/ and css assets in assets/
+  app.use('/assets', express.static(frontendDir));
+  app.engine('mustache', cons.mustache);
+  app.set('view engine', 'mustache');
+  app.set('views', templateDir);
+
+  app.use(oidc.router);
+
+  app.get('/', (req, res) => {
+    res.render('home', {
+      isLoggedIn: !!req.userinfo,
+      userinfo: req.userinfo
+    });
+  });
+
+  app.get('/profile', oidc.ensureAuthenticated(), (req, res) => {
+    // Convert the userinfo object into an attribute array, for rendering with mustache
+    const attributes = Object.entries(req.userinfo);
+    res.render('profile', {
+      isLoggedIn: !!req.userinfo,
+      userinfo: req.userinfo,
+      attributes
+    });
+  });
+
+  app.get('/logout', (req, res) => {
+    req.logout();
+    res.redirect('/');
+  });
+
+  oidc.on('ready', () => {
+    app.listen(sampleConfig.server.port, () => console.log(`App started on port ${sampleConfig.server.port}`));
+  });
+
+  oidc.on('error', err => {
+    // An error occurred while setting up OIDC
+    throw err;
+  });
+};
