+++ common/views/home.mustache
@@ -2,10 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8">
-    <title>App</title>
-    <link href="/assets/css/okta-sign-in.min.css" type="text/css" rel="stylesheet"/>
-    <link href="/assets/css/okta-theme.css" type="text/css" rel="stylesheet"/>
-    <link href="/assets/css/hljs/tomorrow.css" type="text/css" rel="stylesheet"/>
+    <title>Express Sample App</title>
     <link href="/assets/css/semantic.min.css" type="text/css" rel="stylesheet"/>
     <link href="/assets/css/samples.css" type="text/css" rel="stylesheet"/>
     <base href="/"/>
