+++ resource-server/post-install.js
@@ -3,7 +3,9 @@ const createConfigFile = require('../tools/create-config-file');
 const sampleConfig = {
   oktaSample: {
     oidc: {
-      issuer: 'https://{yourOktaDomain}.com/oauth2/default'
+      issuer: 'https://{yourOktaDomain}.com/oauth2/default',
+      aud: 'api://default',
+      cid: '{yourClientId}'
     },
     server: {
       port: 8000
