+++ tools/create-config-file.js
@@ -4,6 +4,7 @@
 
 // This creates a boilerplate config file for you
 
+const colors = require('colors');
 const fs = require('fs');
 const path = require('path');
 
@@ -14,4 +15,5 @@ module.exports = function createConfigFile(sampleConfig) {
     console.log('Creating default configuration file');
     fs.writeFileSync(file, JSON.stringify(sampleConfig, '\n', 2) + '\n');
   }
+  console.log(colors.green(`\nSample project is ready to go!  Please add your configuration to ${file}, see the README for instructions.\n`));
 };
